refactor: split app into components
Add feature-game lib
Add modules and components for board, dice, phase, action, piles, hand and domain
Add ng-mocks as dev dep
Setup minimal passing tests
