+++ apps/taormina-duel/src/app/app.component.spec.ts
@@ -1,41 +1,38 @@
 import { TestBed } from '@angular/core/testing';
-import { FormsModule } from '@angular/forms';
 import { StoreModule } from '@ngrx/store';
-import {
-  DiscardPileCardsFacade,
-  DomainsCardsFacade,
-  EventsPileCardsFacade,
-  FaceUpPilesCardsFacade,
-  GameFacade,
-  HandsCardsFacade,
-  LandsPileCardsFacade,
-  StockPilesCardsFacade,
-} from '@taormina/data-access-game';
 import { GameRulesService } from '@taormina/feature-engine';
+import { FeatureGameModule } from '@taormina/feature-game';
+import { MockInstance, MockModule, MockProvider } from 'ng-mocks';
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       declarations: [AppComponent],
-      imports: [FormsModule, StoreModule.forRoot({})],
-      providers: [
-        GameFacade,
-        DomainsCardsFacade,
-        HandsCardsFacade,
-        FaceUpPilesCardsFacade,
-        LandsPileCardsFacade,
-        StockPilesCardsFacade,
-        EventsPileCardsFacade,
-        DiscardPileCardsFacade,
-        GameRulesService,
-      ],
+      imports: [StoreModule.forRoot({}), MockModule(FeatureGameModule)],
+      providers: [MockProvider(GameRulesService)],
     }).compileComponents();
   });
 
   it('should create the app', () => {
     const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app).toBeTruthy();
+    const component = fixture.componentInstance;
+    expect(component).toBeTruthy();
+  });
+
+  describe('startNewGame', () => {
+    const initNewGame = jest.fn();
+    beforeEach(async () => {
+      MockInstance(GameRulesService, () => ({
+        initNewGame,
+      }));
+      const fixture = TestBed.createComponent(AppComponent);
+      const component = fixture.componentInstance;
+      component.startNewGame();
+    });
+
+    it('should call initNewGame', () => {
+      expect(initNewGame).toHaveBeenCalled();
+    });
   });
 });
