+++ apps/taormina-duel/src/app/app.module.ts
@@ -1,20 +1,22 @@
-import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { DataAccessGameModule } from '@taormina/data-access-game';
-
-import { AppComponent } from './app.component';
-import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
-import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { environment } from '../environments/environment';
-import { FormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
+import { EffectsModule } from '@ngrx/effects';
+import { StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+
+import { DataAccessGameModule } from '@taormina/data-access-game';
+import { FeatureGameModule } from '@taormina/feature-game';
+
+import { environment } from '../environments/environment';
+import { AppComponent } from './app.component';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     BrowserModule,
-    FormsModule,
+    BrowserAnimationsModule,
     StoreModule.forRoot(
       {},
       {
@@ -28,7 +30,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
     EffectsModule.forRoot([]),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
     DataAccessGameModule,
-    BrowserAnimationsModule,
+    FeatureGameModule,
   ],
   providers: [],
   bootstrap: [AppComponent],
