+++ libs/feature-game/README.md
@@ -1,0 +1,7 @@
+# feature-game
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test feature-game` to execute the unit tests.
