+++ libs/feature-game/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/feature-game.module';
