+++ libs/feature-game/src/lib/actions/actions.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ActionsComponent } from './actions/actions.component';
+
+@NgModule({
+  declarations: [ActionsComponent],
+  imports: [CommonModule],
+  exports: [ActionsComponent],
+})
+export class ActionsModule {}
