+++ libs/feature-game/src/lib/actions/actions/actions.component.css
@@ -1,0 +1,4 @@
+.flex-container {
+  display: flex;
+  column-gap: 20px;
+}
