+++ libs/feature-game/src/lib/board/board/board.component.html
@@ -1,0 +1,15 @@
+<taormina-phases></taormina-phases>
+<hr />
+<taormina-dice></taormina-dice>
+<hr />
+<taormina-actions></taormina-actions>
+<hr />
+<taormina-piles></taormina-piles>
+<hr />
+<taormina-hand [handId]="ID_HAND_RED"></taormina-hand>
+<hr />
+<taormina-domain [domainId]="ID_DOMAIN_RED"></taormina-domain>
+<hr />
+<taormina-domain [domainId]="ID_DOMAIN_BLUE"></taormina-domain>
+<hr />
+<taormina-hand [handId]="ID_HAND_BLUE"></taormina-hand>
