+++ libs/feature-game/src/lib/domain/domain/domain.component.css
@@ -1,8 +1,3 @@
-.flex-container {
-  display: flex;
-  column-gap: 20px;
-}
-
 .grid-container {
   display: grid;
   grid-gap: 20px;
