+++ libs/feature-game/src/lib/domain/domain/domain.component.html
@@ -1,0 +1,91 @@
+<div *ngIf="getDomain() as domain">
+  <div>
+    <span>{{ domain.color }}</span>
+    <span> - Ships count: {{ getShipsCount(domain.id) | async }}</span>
+    <span>
+      - Celebration points: {{ getCelebrationPoints(domain.id) | async }}</span
+    >
+    <span *ngIf="getTradeMastery(domain.id) | async"> - Trade Mastery</span>
+    <span *ngIf="getStrengthMastery(domain.id) | async">
+      - Strength Mastery</span
+    >
+    <span> - Victory points: {{ getVictoryPoints(domain.id) | async }}</span>
+  </div>
+  <div
+    class="grid-container"
+    [ngStyle]="{
+      'grid-template-columns': getColumnsTemplate(domain.id) | async,
+      'grid-template-rows': getRowsTemplate(domain.id) | async
+    }"
+  >
+    <div
+      *ngFor="let pivot of getDomainsCards(domain.id) | async"
+      [ngSwitch]="pivot.cardType"
+      class="grid-cell"
+      [ngStyle]="{
+        'grid-column-start': 'col' + pivot.col,
+        'grid-row-start': 'row' + pivot.row
+      }"
+    >
+      <div *ngSwitchCase="DEVELOPMENT_CARD_INTERFACE_NAME">
+        <div
+          *ngIf="
+            pivot.cardId !== undefined &&
+            getDevelopmentCard(pivot.cardId) as development
+          "
+        >
+          {{ development.name }}
+        </div>
+      </div>
+      <div *ngSwitchCase="AGGLOMERATION_CARD_INTERFACE_NAME">
+        <div
+          *ngIf="
+            pivot.cardId !== undefined &&
+            getAgglomerationCard(pivot.cardId) as agglomeration
+          "
+        >
+          {{ agglomeration.type }}
+          <button type="button" (click)="toggleDomainCardSelection(pivot.id)">
+            {{
+              (isDomainCardSelected(pivot.id) | async) ? 'Unselect' : 'Select'
+            }}
+          </button>
+        </div>
+      </div>
+      <div *ngSwitchCase="LAND_CARD_INTERFACE_NAME">
+        <div
+          *ngIf="
+            pivot.cardId !== undefined && getLandCard(pivot.cardId) as land
+          "
+        >
+          {{ land.type }}<br />
+          {{ land.die }}<br />
+          Available: {{ pivot.availableResources }}<br />
+          locked: {{ pivot.lockedResources }}<br />
+          <button type="button" (click)="lockResource(pivot.id)">Lock</button>
+          <button type="button" (click)="unlockResources(pivot.id)">
+            Unlock</button
+          ><br />
+          <button type="button" (click)="increaseResources(pivot.id)">
+            Increase
+          </button>
+        </div>
+      </div>
+      <div *ngSwitchCase="AVAILABLE_AGGLOMERATION_SLOT">
+        <button type="button" (click)="toggleDomainCardSelection(pivot.id)">
+          {{ (isDomainCardSelected(pivot.id) | async) ? 'Unselect' : 'Select' }}
+        </button>
+      </div>
+      <div *ngSwitchCase="AVAILABLE_DEVELOPMENT_SLOT">
+        <button type="button" (click)="toggleDomainCardSelection(pivot.id)">
+          {{ (isDomainCardSelected(pivot.id) | async) ? 'Unselect' : 'Select' }}
+        </button>
+      </div>
+      <div *ngSwitchCase="AVAILABLE_LAND_SLOT">
+        <button type="button" (click)="toggleDomainCardSelection(pivot.id)">
+          {{ (isDomainCardSelected(pivot.id) | async) ? 'Unselect' : 'Select' }}
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
