+++ libs/feature-game/src/lib/feature-game.module.ts
@@ -1,0 +1,9 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { BoardModule } from './board/board.module';
+
+@NgModule({
+  imports: [CommonModule, BoardModule],
+  exports: [BoardModule],
+})
+export class FeatureGameModule {}
