+++ libs/feature-game/src/lib/phases/phases/phases.component.css
@@ -1,0 +1,4 @@
+.flex-container {
+  display: flex;
+  column-gap: 20px;
+}
