+++ libs/feature-game/src/lib/piles/piles/piles.component.css
@@ -1,0 +1,4 @@
+.flex-container {
+  display: flex;
+  column-gap: 20px;
+}
