+++ package.json
@@ -107,6 +107,7 @@
     "jest-preset-angular": "11.1.2",
     "leasot": "^12.0.0",
     "lint-staged": "^12.1.2",
+    "ng-mocks": "^14.0.1",
     "nx": "14.4.0",
     "prettier": "2.7.1",
     "rxjs-marbles": "^6.0.1",
