+++ tsconfig.base.json
@@ -17,6 +17,7 @@
     "paths": {
       "@taormina/data-access-game": ["libs/data-access-game/src/index.ts"],
       "@taormina/feature-engine": ["libs/feature-engine/src/index.ts"],
+      "@taormina/feature-game": ["libs/feature-game/src/index.ts"],
       "@taormina/shared-constants": ["libs/shared-constants/src/index.ts"],
       "@taormina/shared-models": ["libs/shared-models/src/index.ts"],
       "@taormina/shared-utils": ["libs/shared-utils/src/index.ts"],
